requirements.txt
llama-index-llms-openai
llama-index-readers-file
pymupdf
llama-index
python-dotenv
